[add] Controllerの追加

+++ src/main/java/com/example/springboottodo/repository/TodoRepository.java
@@ -5,4 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TodoRepository extends JpaRepository<Todo, Integer> {}
+public interface TodoRepository extends JpaRepository<Todo, Long> {}
